--- app/grouping/feedback/FeedbackFormProvider.tsx
@@ -1,10 +1,0 @@
-'use client';
-import { FormProvider, useForm } from 'react-hook-form';
-
-import type { StrictPropsWithChildren } from '@/types';
-
-export default function FeedbackFormProvider({ children }: StrictPropsWithChildren) {
-  const methods = useForm();
-
-  return <FormProvider {...methods}>{children}</FormProvider>;
-}
