+++ app/grouping/feedback/components/FeedbackWrapper.tsx
@@ -1,10 +1,18 @@
 'use client';
 import MateComponent from './mate/MateComponent';
 import PraiseComponent from './praise/PraiseComponent';
+import { useFeedbackContext } from '../FeedbackContext';
 import { useFunnel } from '@/hooks/useFunnel';
+
+import type { FeedbackRequestType } from '../type';
 
 export default function FeedbackWrapper() {
   const { Funnel, prevStep, nextStep } = useFunnel(['praise', 'mate']);
+  const { handleSubmit, control } = useFeedbackContext();
+
+  const onSubmit = (data: FeedbackRequestType) => {
+    console.log(data);
+  };
 
   return (
     <Funnel>
@@ -12,7 +20,11 @@ export default function FeedbackWrapper() {
         <PraiseComponent onPrevClick={prevStep} onNextClick={nextStep} />
       </Funnel.Step>
       <Funnel.Step name="mate">
-        <MateComponent onPrevClick={prevStep} onNextClick={nextStep} />
+        <MateComponent
+          onPrevClick={prevStep}
+          onNextClick={handleSubmit(onSubmit)}
+          control={control}
+        />
       </Funnel.Step>
     </Funnel>
   );
